Get a working (but slow) hwacha-flat
I've only tested this on Spike, and it uses way too many registers.  I
think another good thing to do would be to precompute '(hot[i] -
cold[i]) * 255.0f', as that would save a significant number of
instructions.

That said, this does at least work, so I'm going to try it on the
FPGA.